Cast to a specific destination view controller subclass
Introduces a generic type parameter to `Segue`, which is used to perform
a dynamic cast on the segue's destination view controller.

Makeing `perform(_:configure:)` generic had the unexpected effect of
making this call ambiguous to the type checker:

    root?.perform(.showDetails) {
      // ...
    }

The fix is to introduce an overload of `perform`, which somehow makes
the generic method non-ambiguous, even if the overload is marked as
unavailable.