Disable App Extension API
It'd be nice to be able to do this (and _our_ code does only use the app
extension API), but since we're linking in Aspects (which _doesn't_ only use
the app extension API), we actually can't make that guarantee. We should
disable it to avoid the resulting warnings as well as any confusion that we
might run into.